requirements.txt
Flask==2.3.0
rembg==2.0.30
Pillow==9.5.0
